Fix syntax errors in view/conversions.h (#43537)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43537

This file always had a bunch of errors and made it so that actual errors do not show up (too many emitted). It was just because we are not including RawProps.h. I am not sure how it ever got compiled but I guess some header included something that included something etc. Maybe its a vs code issue but this seems like an obvious fix

Changelog: [Internal]

Reviewed By: NickGerleman

Differential Revision: D55041883

fbshipit-source-id: 3445770dd25dbe294024649f96f9e3af7777b2b2